refactor(game): extract auto-start IDLE effect to a hook (#184)
Second slice of #116 (after #183 extracted keyboard shortcuts).
GameContainer drops further from 652 → 644 LOC.

What moved
==========

src/hooks/useAutoStartIdle.ts — new file, 37 lines. The mount-time
effect that:

  1. Auto-starts a new game when status is IDLE, honouring deep-link
     ?type= and ?difficulty= query params (validated against the
     constants tables — invalid values fall back to current state).
  2. For a returning user mid-game (PLAYING / PAUSED), records a
     game-start event so a completion this session shows up in stats.
     Skips COMPLETED / LOST to avoid inflating gamesStarted on every
     reload of a finished game.

Empty-deps + eslint-disable preserved verbatim — the effect is
explicitly mount-only by design (subsequent state changes go through
handleNewGame and friends).

GameContainer becomes
=====================

  useAutoStartIdle(state, actions);

…instead of an inline 17-line useEffect. The DifficultyLevel /
SudokuTypeId / SUDOKU_TYPES imports moved with the effect; the
constants imports the rest of the file still uses (DIFFICULTY_LEVELS,
GAME_STATUS, EMPTY_CELL) stay.

Drive-by lint fix
=================

The `react-hooks/refs` rule now flags the inline ref-mirror at line
87 (`elapsedTimeRef.current = state.elapsedTime`) on this branch but
not on dev — almost certainly an eslint-plugin-react-hooks 7.0.1
context-sensitive rule firing differently after my edit shifted line
numbers. Added a targeted `eslint-disable-next-line react-hooks/refs`
with explanatory comment. The pattern itself is the documented
workaround for "read latest value in effect without re-running it"
until React's useEffectEvent stabilizes.

Verified
========

- typecheck: 0 errors
- lint: clean
- Full unit suite passes
- GameContainer.tsx: 652 → 644 LOC (-8 net; the inline 17-line
  effect was replaced with a 1-line hook call + 2 lines of inline
  comment, and the unused-now imports were removed)
- New file: 37 LOC
